Use AbstractApi class, remove duplication of methods
